# app/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Storefront API
  version: 1.0.0
  description: |
    REST API for the storefront backend — accounts, catalog, cart, checkout, orders,
    shipping addresses and an admin namespace.

    ## Base path

    Every API route is served under `/api/v1`, e.g. `POST /api/v1/auth/login`. The prose
    below leaves the prefix out. Only `GET /` (health check), `GET /docs` and
    `GET /openapi.yaml` sit at the server root.

    ## Response envelope

    Every route except `GET /` wraps its payload in the same envelope:

    ```json
    { "status": 200, "message": "Products fetched.", "data": [] }
    ```

    `status` repeats the HTTP status code and `data` is `null` on errors. An error adds a
    machine-readable `code`:

    ```json
    { "status": 401, "message": "Access token has expired.", "data": null, "code": "token_expired" }
    ```

    `code` is one of `invalid_request`, `bad_request`, `no_token`, `token_expired`,
    `token_invalid`, `invalid_credentials`, `forbidden`, `not_found`, `conflict`,
    `rate_limited` or `internal_error`.

    ## Pagination

    Lists that can grow — products, orders and the admin lists — take `limit`
    (1–100, default 50) and `offset` (default 0), and add a `meta` object next to `data`:

    ```json
    { "status": 200, "message": "Products fetched.", "data": [], "meta": { "limit": 50, "offset": 0, "total": 134 } }
    ```

    ## Authentication

...

    ```
    Authorization: Bearer <accessToken>
    ```

...

...

    ```
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/users/{id}/password:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [Users]
      summary: Change your password
      description: |
        Takes the current password as well as the new one, so a stolen access token alone
        cannot lock the owner out. Every other session of the account ends, and the reply
        carries a fresh access token and refresh cookie so this client stays signed in.
      operationId: changeOwnPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password, example: pass1234 }
                newPassword:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  example: newpass1234
      responses:
        '200':
          description: The password was changed and this session renewed.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/RefreshCookie'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Password changed. Other sessions were signed out.
                      data: { $ref: '#/components/schemas/AuthSession' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: No token, or the current password is wrong.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example:
                status: 401
                message: Current password is incorrect
                data: null
                code: invalid_credentials
        '403':
          $ref: '#/components/responses/NotYourAccount'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Products ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The `accessToken` from `POST /auth/register`, `POST /auth/login` or
        `POST /auth/refresh`, sent as `Authorization: Bearer <accessToken>`.
        Expires after 15 minutes by default. The refresh token is not sent this way: it
        travels in an HttpOnly cookie.
  headers:
    RefreshCookie:
      description: |
        The refresh token, in a cookie JavaScript cannot read:
        `HttpOnly`, `SameSite=Strict`, `Path=/api/v1/auth`, and `Secure` in production.
      schema: { type: string }
      example: refreshToken=6f1a…; Max-Age=604800; Path=/api/v1/auth; HttpOnly; SameSite=Strict
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: User id.
      schema: { type: integer, minimum: 1 }
      example: 1
...
  schemas:
    Envelope:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: integer
          description: Repeats the HTTP status code.
          example: 200
        message:
          type: string
          description: A sentence a client may show as-is.
          example: Products fetched.
        data:
          description: The payload, or `null` on an error.
          nullable: true
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          required: [code]
          properties:
            data: { type: object, nullable: true, example: null }
            code:
              type: string
              description: What went wrong, in a form a client can branch on.
              enum:
                - invalid_request
                - bad_request
                - no_token
                - token_expired
                - token_invalid
                - invalid_credentials
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - internal_error
...
    UserRole:
      type: string
      enum: [customer, admin]
      example: customer
...
    User:
      type: object
      description: No password field exists on this shape, so a hash cannot leak by accident.
      properties:
        id: { type: integer, example: 1 }
        firstName: { type: string, example: John }
        lastName: { type: string, example: Doe }
        username: { type: string, example: johndoe }
        role: { $ref: '#/components/schemas/UserRole' }
...
    AuthSession:
      type: object
      description: |
        What a session start or renewal returns. The refresh token is not in here — it is set
        as an HttpOnly cookie.
      properties:
        user: { $ref: '#/components/schemas/User' }
        accessToken:
          type: string
          description: HS256 JWT, valid for 15 minutes by default.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
  responses:
    BadRequest:
      description: The request does not add up. `message` names the field that was wrong.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: 400
            message: quantity must be a whole number between 1 and 999
            data: null
            code: invalid_request
...
    NotYourAccount:
      description: The id belongs to another account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: 403
            message: You can only access your own data
            data: null
            code: forbidden
...
    NotFound:
      description: Nothing with that id.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: 404
            message: product with id 999 not found
            data: null
            code: not_found
...
    InternalError:
      description: |
        Something went wrong on the server. The real error is logged with the request id from
        the `X-Request-Id` header; the answer stays deliberately vague.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: 500
            message: Internal Server Error
            data: null
            code: internal_error